pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "stamp-photo"
version = "1.0.0"
description = "Postage stamp style photo generator / 邮票风格照片生成器"
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "orangentle", url = "https://github.com/orangentle" },
]
dependencies = [
    "Pillow>=9.0",
    "PySide6>=6.4",
]

[project.urls]
Homepage = "https://github.com/orangentle/stamp-photo"
Repository = "https://github.com/orangentle/stamp-photo"
Issues = "https://github.com/orangentle/stamp-photo/issues"

[project.gui-scripts]
stamp-photo = "stamp_photo.gui:main"
